Resolve merge conflicts with main branch
- Fixed import issues by replacing WorkspaceHash with ProjectRootId
- Updated database schema references from workspace_id to project_root_path
- Preserved all workspace ID repository root functionality
- All code now compiles successfully

+++ crates/forge_app/src/tool_executor.rs
@@ -208,7 +208,7 @@ impl<
                             &search_query.query,
                             &search_query.use_case,
                         )
-                        .limit(limit)
+                        .limit(limit as usize)
                         .top_k(top_k);
                         if let Some(ext) = &input.file_extension {
                             params = params.ends_with(ext);
